add auth guard in the layout component. add service to verify if the token exist. update request service, now process the token errors. update alert message errors in the components

+++ src/app/components/navbar/navbar.component.ts
@@ -13,8 +13,10 @@ export class NavbarComponent {
   constructor(private router: Router, private cookieService: CookieService){}
 
   closeSesion(){
-    localStorage.removeItem('token');
-    this.cookieService.delete('user');
-    this.router.navigate(["/auth/login"]);
+    if(window.confirm("Estas seguro que quieres cerrar sesion?")){
+      localStorage.removeItem("token");
+      this.cookieService.delete("user");
+      this.router.navigate(["/auth/login"]);
+    }
   } 
 }
